Refactor reddit meme to generator
